requirements.txt
emoji
colorama
motor
dnspython
requests
discord.py